l900: Fixes to Extract Files Script
This commit fixes a path bug in the extract-files.sh script.

The l900 extract-files.sh includes both the t0lte's
extract-files.sh and the smdk4412-common's extract-files.sh;
this is where the majority of the blobs are pulled in, as
l900's proprietary-files.txt only notes 3 of its own.

However, due to a bug in the includes in this script, when
vendor makefiles are generated for t0lte & smdk4412-common
as part of running the l900 script, the makefiles use the
proprietary-files.txt from the l900 folder instead of their
own.

This results in only the 3 files from l900's list  being copied
to the  device, and a build to be missing a lot of necessary
binary blobs. Among other things, this results in the device
having no radio.

Additionally, there are downstream issues in t0lte - it seems
devices like the l900 are built out of common.mk, where t0lte.mk
can also be built directly. l900 includes common.mk, which
expects to include t0lte-common-vendor.mk - however,
extract-files.sh creates t0lte-vendor.mk, which t0lte.mk expects.
There is no way for the script to create t0lte-common-vendor.mk.

A seperate patch will be submitted for t0lte to fix this behavior.
by introducing a parameter to indicate which set to generate.

With both patches applied, a working build can be made for t0lte
using extract-files.sh to pull blobs.

Change-Id: I9e4a665640c809f69aede73d9f59f200f5c8f494